fix: harden websocket permessage-deflate streaming and frame parser
decompress_chunk infinite-looped and corrupted multi-chunk messages: when a
read or fragment boundary fell mid deflate-block, inflate left the partial
code unconsumed and the loop spun forever. Even when broken out of, the
unconsumed bytes were discarded each call (compressed_buf overwritten from
offset 0), corrupting the stream across chunks/fragments.

- retain unconsumed inflate input at the front of compressed_buf
  (compressed_buffered/compressed_consumed), compact the consumed prefix,
  re-derive next_in after ensure_capacity reallocs; reset only between
  messages, never between fragments
- break the inflate loop on no progress instead of spinning
- bound the decompressed size by client_max_body_size (bomb guard)
- reject RSV1 on continuation frames (RFC 7692 §7.2)
- free compressed_buf on parser free (leak)
- drop dead mask_index field
- use ssize_t for connection_data_read's result in the read loop

+++ protocols/websocket/server/websocketsserverhandlers.c
@@ -51,7 +51,7 @@ int __read(connection_t* connection) {
     websocketsparser_t* parser = ctx->parser;
 
     while (1) {
-        int bytes_readed = 0;
+        ssize_t bytes_readed = 0;
         read_data:
 
         bytes_readed = connection_data_read(connection);
